feat(bet4-sandbox): extract sandbox tools into a scoped ToolPack (EP-8DC217EB BET-4)
13 sandbox tools (check/diagnose/edit/iterate/list/read/recover/run_command/run_tests/search/start/write_sandbox + get_build_sandbox_state) moved behavior-identically into sandbox-pack.ts; build-helper users import from build-tool-helpers; the generate_code/iterate_sandbox shared removed-guard is dropped (both are handler-only in their packs now). Baseline 16, tsc 0, full web vitest green. Collapsed to one signed commit to satisfy DCO.
