+++ scripts/mcp-tool-pack-baseline.json
@@ -1,31 +1,18 @@
 [
-  "check_sandbox",
   "create_build_epic",
   "create_portal_pr",
   "deploy_feature",
-  "diagnose_sandbox",
-  "edit_sandbox_file",
-  "get_build_sandbox_state",
-  "iterate_sandbox",
-  "list_sandbox_files",
   "principle_decide",
-  "read_sandbox_file",
-  "recover_sandbox",
   "register_digital_product_from_build",
   "reviewBuildPlan",
   "reviewDesignDoc",
-  "run_sandbox_command",
-  "run_sandbox_tests",
   "run_ux_test",
   "saveBuildEvidence",
-  "search_sandbox",
   "set_change_disposition",
   "start_build",
   "start_ideate_research",
-  "start_sandbox",
   "start_scout_research",
   "update_feature_brief",
   "validate_schema",
-  "verification_preflight",
-  "write_sandbox_file"
+  "verification_preflight"
 ]
