+++ scripts/module-size-baseline.txt
@@ -40,6 +40,9 @@ apps/web/lib/integrate/build-orchestrator.ts	1780
 apps/web/lib/integrate/sandbox/build-branch.ts	854
 apps/web/lib/marketing.ts	1367
 apps/web/lib/mcp-tools-backlog.test.ts	901
+apps/web/lib/mcp-tools.ts	7354
+apps/web/lib/mcp/packs/backlog-pack.ts	1380
+apps/web/lib/mcp/packs/sandbox-pack.ts	920
 apps/web/lib/mcp-tools.ts	9860
 apps/web/lib/mcp/packs/contribution-hive-pack.ts	902
 apps/web/lib/mcp-tools.ts	8633
